MainForm class: changed: speed of Body, time to YMDH,  deleted TransferMassCenter.
Scale class: changed Time to int.
Canvas class: added TransferMassCenter; changed SecondsFromStart and Center.

+++ MotionSimulation/Scale.cs
@@ -9,10 +9,10 @@ namespace MotionSimulation
     public class Scale
     {
         public double Length { get; set; }
-        public double Time { get; set; }
+        public int Time { get; set; }
         public double Radius { get; set; }
 
-        public Scale(double length, double time, double radius = 1)
+        public Scale(double length, int time, double radius = 1)
         {
             Length = length;
             Time = time;
